+++ scanpointgenerator/core/dimension.py
@@ -45,12 +45,31 @@ class Dimension(object):
         Returns:
             Positions (np.array): Array of positions
         """
+        # check that this dimension is prepared
+        if not self._prepared:
+            raise ValueError("Must call prepare first")
+        # get the mesh map for this axis and use it to get the generator points
+        gen = [g for g in self.generators if axis in g.axes][0]
+        return gen.positions[axis][self.get_mesh_map(axis)]
+
+    def get_mesh_map(self, axis):
+        """
+        Retrieve the mesh map (indices) for a given axis within the dimension.
+
+        Args:
+            axis (str): axis to get positions for
+        Returns:
+            Positions (np.array): Array of mesh indices
+        """
         # the points for this axis must be scaled and then indexed
         if not self._prepared:
             raise ValueError("Must call prepare first")
         # scale up points for axis
         gen = [g for g in self.generators if axis in g.axes][0]
         points = gen.positions[axis]
+        # just get index of points instead of actual point value
+        points = np.arange(len(points))
+
         if self.alternate:
             points = np.append(points, points[::-1])
         tile = 0.5 if self.alternate else 1
@@ -66,7 +85,6 @@ class Dimension(object):
         else:
             points = np.tile(points, int(tile))
         return points[self.indices]
-
 
     def apply_excluder(self, excluder):
         """Apply an excluder with axes matching some axes in the dimension to